+++ src/aks-preview/HISTORY.rst
@@ -9,13 +9,24 @@ If there is no rush to release a new version, please just add a description of t
 
 To release a new version, please select a new version number (usually plus 1 to last patch version, X.Y.Z -> Major.Minor.Patch, more details in `\doc <https://semver.org/>`_), and then add a new section named as the new version number in this file, the content should include the new modifications and everything from the *Pending* section. Finally, update the `VERSION` variable in `setup.py` with this new version number.
 
-19.0.0b4
+19.0.0b6
 +++++++
 * `az aks nodepool get-rollback-versions`: Add new command to get available rollback versions for a nodepool.
 * `az aks nodepool rollback`: Add new command to rollback a nodepool to a previously used configuration (N-1).
 
 Pending
 +++++++
+
+19.0.0b5
++++++++
+* `az aks get-credentials`: Convert device code mode kubeconfig to Azure CLI token format to bypass conditional access login blocks.
+* Add `enable-istio-cni` and `disable-istio-cni` commands under `az aks mesh`.
+
+19.0.0b4
++++++++
+* `az aks create`: Add new parameter `--enable-hosted-system` to enable hosted system components for automatic clusters.
+* Added `--enable-gateway-api` to `az aks create` to enable managed Gateway API installation
+* Added `--enable-gateway-api` and `--disable-gateway-api` to `az aks update` to enable/disable managed Gateway API installation
 
 19.0.0b3
 +++++++